Librarian: Add Location to Item Details, Enhance UI to reduce confusion
